requirements.txt
flask
flask_login
supabase
python-dotenv
werkzeug